# openapi.json
{
...
  "info": {
    "description": "Auto-generated route inventory for this iterion instance. Routes and methods are exact; request/response schemas are populated for the typed surface (see components) and enriched route-by-route. Note: the /api/v1/native, /api/v1/dispatcher and /api/v1/mcp/board sub-trees are served but registered on a separate mux, so they are intentionally not catalogued here.",
    "title": "iterion API",
    "version": "current"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/runs/{id}": {
      "delete": {
        "operationId": "deleteRunsById",
        "responses": {
          "default": {
            "description": "Response"
          }
        },
        "summary": "DELETE /api/runs/{id}",
        "tags": [
          "runs"
        ]
      },
      "get": {
        "operationId": "getRunsById",
        "responses": {
          "default": {
            "description": "Response"
          }
        },
        "summary": "GET /api/runs/{id}",
        "tags": [
          "runs"
        ]
      },
      "parameters": [
        {
          "in": "path",
          "name": "id",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ]
    },
...
    "/api/ws/runs/{id}": {
      "get": {
        "operationId": "getWsRunsById",
        "responses": {
          "default": {
            "description": "Response"
          }
        },
        "summary": "GET /api/ws/runs/{id}",
        "tags": [
          "ws"
        ]
      },
      "parameters": [
        {
          "in": "path",
          "name": "id",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ]
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ]
}